+++ src/settings/Settings.ts
@@ -25,6 +25,7 @@ export interface SemesterSettings {
     topicColors: Record<string, string>;
 
     showBoard: boolean;
+    showTaskActions: boolean;
     hasSeenWelcome: boolean;
     hasClickedRibbonIcon: boolean;
 
@@ -60,6 +61,7 @@ export const DEFAULT_SETTINGS: SemesterSettings = {
     appWideAutomation: true,
 
     showBoard: false,
+    showTaskActions: false,
     hasSeenWelcome: false,
     hasClickedRibbonIcon: false,